Add files via upload
We want to visit 45 cities over 45 days. On each day, we visit different city by plane. Note that the price of plane tickets can differ on each day. We start and end our trip in Prague. Goal: Find the route with the cheapest plane tickets. Or at least find a route with a reasonable cost. Or at least find a route with available flights.
Solution was found using simulations in R. Code contains different simulation algorithms that were tried in order to find the best one.